// demo/Blackholio/client-ts/src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Circle = __t.object("Circle", {
  entityId: __t.i32(),
  playerId: __t.i32(),
  get direction() {
    return DbVector2;
  },
  speed: __t.f32(),
  lastSplitTime: __t.timestamp(),
});
export type Circle = __Infer<typeof Circle>;

export const CircleDecayTimer = __t.object("CircleDecayTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type CircleDecayTimer = __Infer<typeof CircleDecayTimer>;

export const CircleRecombineTimer = __t.object("CircleRecombineTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
  playerId: __t.i32(),
});
export type CircleRecombineTimer = __Infer<typeof CircleRecombineTimer>;

export const Config = __t.object("Config", {
  id: __t.i32(),
  worldSize: __t.i64(),
});
export type Config = __Infer<typeof Config>;

export const ConsumeEntityEvent = __t.object("ConsumeEntityEvent", {
  consumedEntityId: __t.i32(),
  consumerEntityId: __t.i32(),
});
export type ConsumeEntityEvent = __Infer<typeof ConsumeEntityEvent>;

export const ConsumeEntityTimer = __t.object("ConsumeEntityTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
  consumedEntityId: __t.i32(),
  consumerEntityId: __t.i32(),
});
export type ConsumeEntityTimer = __Infer<typeof ConsumeEntityTimer>;

export const DbVector2 = __t.object("DbVector2", {
  x: __t.f32(),
  y: __t.f32(),
});
export type DbVector2 = __Infer<typeof DbVector2>;

export const Entity = __t.object("Entity", {
  entityId: __t.i32(),
  get position() {
    return DbVector2;
  },
  mass: __t.i32(),
});
export type Entity = __Infer<typeof Entity>;

export const Food = __t.object("Food", {
  entityId: __t.i32(),
});
export type Food = __Infer<typeof Food>;

export const MoveAllPlayersTimer = __t.object("MoveAllPlayersTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type MoveAllPlayersTimer = __Infer<typeof MoveAllPlayersTimer>;

export const Player = __t.object("Player", {
  identity: __t.identity(),
  playerId: __t.i32(),
  name: __t.string(),
});
export type Player = __Infer<typeof Player>;

export const SpawnFoodTimer = __t.object("SpawnFoodTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type SpawnFoodTimer = __Infer<typeof SpawnFoodTimer>;